Update installation.rst
This fixes 2 typos in the name PROJECTS_FOLDER, and makes clear the fact that `<PROJECTS_FOLDER>` is a symbolic reference the user needs to replace with their actual pathname.
It does not yet, but should be further modified to provide an example of what "of the form `/<DRIVE_LETTER>/<PATH>`" actually looks like. (for example, if the drive letter is C, does it need to include the colon after `C:` or not?)
It also does not yet, but should give an example of what the full command looks like, when "adding `/app/slybot/bin/portiacrawl <PROJECT_PATH> [SPIDER] [OPTIONS]` to the command above"
It also does not yet, but should explain what is meant by PROJECT_PATH, SPIDER, AND OPTIONS, since they are also symbolic references the user needs to replace with meaningful strings. If this is truly the install instructions read by a first time user, the user might not know what any of these things are yet.